+++ packages/importer/src/utils.ts
@@ -1,13 +1,6 @@
 import { transliterate } from 'transliteration';
 import camelcase from 'camelcase';
 
-export function splitCommas(value: unknown): string[] | undefined {
-  if (!value) return;
-  return String(value)
-    .split(',')
-    .map((s) => s.trim());
-}
-
 export function normalizeColumnName(value: string) {
   const parts = value.split('.');
   return parts.map((s) => camelcase(transliterate(s)).replace(/\W/g, '')).join('.');
@@ -58,7 +51,7 @@ export function guessType(value: string) {
     return 'float';
   } else if (['true', 'false'].includes(value)) {
     return 'bool';
-  } else if (value.match(/\S+@\S+.\S+/)) {
+  } else if (value.match(/^\S+@\S+\.\S+$/)) {
     return 'email';
   } else if (parseArray(value)) {
     return 'multiple';